+++ aiobml/__init__.py
@@ -1,2 +1,2 @@
 from .client import asyncBML
-__version__ = '0.1.5'
+__version__ = '1.0.0'