+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html);
 
+# v1.0.0
+
+## Added
+
+- New feature: `event` added
+  - use decorators
+  ```py
+  @bank.event('new_transaction')
+	async def on_new_transaction(transaction):
+    print(transaction)
+  ```
+  to be notified of the transactions.
+
+- No breaking changes
+
 # v0.1.5
 
 ### Internal